requirements.txt
# Speech Recognition
SpeechRecognition>=3.10.0
PyAudio>=0.2.13
# Text-to-Speech
gTTS>=2.4.0
# Environment Variables
python-dotenv>=1.0.0
